Extract a condition chain to aiobindings
