+++ kopf/structs/primitives.py
@@ -5,27 +5,9 @@ import asyncio
 import enum
 import threading
 import time
-from typing import Awaitable, Generator, Generic, Optional, TypeVar, Union
+from typing import Awaitable, Generator, Generic, Optional, TypeVar
 
 from kopf.aiokits import aiotasks
-
-
-async def condition_chain(
-        source: asyncio.Condition,
-        target: asyncio.Condition,
-) -> None:
-    """
-    A condition chain is a "clean" hack to attach one condition to another.
-
-    It is a "clean" (not "dirty") hack to wake up the webhook configuration
-    managers when either the resources are revised (as seen in the insights),
-    or a new client config is yielded from the webhook server.
-    """
-    async with source:
-        while True:
-            await source.wait()
-            async with target:
-                target.notify_all()
 
 
 FlagReasonT = TypeVar('FlagReasonT', bound=enum.Flag)
